requirements.txt
# Core
langchain-ollama==1.0.1
langchain==1.2.9
langchain-community==0.4.1
langchain-core==1.2.9
langchain-huggingface==1.2.0

# --- Evaluation & Data ---
ragas==0.4.3
sentence-transformers==5.2.2
datasets==4.5.0
pandas==3.0.0
numpy==1.26.4

# --- Async Utilities ---
nest_asyncio==1.6.0

# --- PDF Processing ---
pdfplumber==0.11.2

# --- Vector Store ---
faiss-cpu==1.9.0

# --- Ignore the entire results folder ---
results/

# --- Ignore specific file types anywhere in the project ---
*.xlsx
*.csv